move Logic filters
