+++ src/Filter/Logic/OrX.php
@@ -7,9 +7,12 @@
  * @license   http://opensource.org/licenses/MIT
  */
 
-namespace Happyr\DoctrineSpecification\Filter;
+namespace Happyr\DoctrineSpecification\Filter\Logic;
 
-class FilterCollection implements Filter
+
+use Happyr\DoctrineSpecification\Filter\Filter;
+
+class OrX implements Filter
 {
     /**
      * @var Filter[]
@@ -17,7 +20,7 @@ class FilterCollection implements Filter
     private $filters;
 
     /**
-     * Construct it with two or more instances of Filter.
+     * Construct it with two or more instances of Specification.
      *
      * @param Filter $filter1
      * @param Filter $filter2
@@ -25,14 +28,24 @@ class FilterCollection implements Filter
     public function __construct(Filter $filter1, Filter $filter2)
     {
         foreach (func_get_args() as $filter) {
-            $this->addFilter($filter);
+            $this->orX($filter);
         }
     }
 
     /**
+     * Append an other specification with a logic OR.
+     *
+     * <code>
+     * $spec = Spec::orX(A, B);
+     * $spec->orX(C);
+     *
+     * // We be the same as
+     * $spec = Spec::orX(A, B, C);
+     * </code>
+     *
      * @param Filter $filter
      */
-    public function addFilter(Filter $filter)
+    public function orX(Filter $filter)
     {
         $this->filters[] = $filter;
     }
