--- src/Transformer/Doctrine/ORM/QueryBuilder/Logic/AndXTransformer.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * Doctrine Specification.
- *
- * @author    Tobias Nyholm <tobias@happyr.com>
- * @copyright Copyright (c) 2014, Tobias Nyholm
- * @license   http://opensource.org/licenses/MIT
- */
-
-namespace Happyr\DoctrineSpecification\Transformer\Doctrine\ORM\QueryBuilder\Logic;
-
-use Doctrine\ORM\QueryBuilder;
-use Happyr\DoctrineSpecification\Logic\AndX;
-use Happyr\DoctrineSpecification\Specification;
-use Happyr\DoctrineSpecification\Transformer\Doctrine\ORM\QueryBuilder\QueryBuilderTransformerCollection;
-use Happyr\DoctrineSpecification\Transformer\Doctrine\ORM\QueryBuilder\QueryBuilderTransformerCollectionAware;
-use Happyr\DoctrineSpecification\Transformer\Doctrine\ORM\QueryBuilder\QueryBuilderTransformerCollectionAwareTrait;
-
-class AndXTransformer implements QueryBuilderTransformerCollectionAware
-{
-    use QueryBuilderTransformerCollectionAwareTrait;
-
-    /**
-     * @param Specification $specification
-     * @param QueryBuilder  $qb
-     * @param string        $dqlAlias
-     *
-     * @return QueryBuilder
-     */
-    public function transform(Specification $specification, QueryBuilder $qb, $dqlAlias)
-    {
-        if ($specification instanceof AndX && $this->collection instanceof QueryBuilderTransformerCollection) {
-            foreach ($specification->getChildren() as $child) {
-                $qb = $this->collection->transform($child, $qb, $dqlAlias);
-            }
-        }
-
-        return $qb;
-    }
-}
